refactor(packages): consolidate remote-data layer — phase 4
Merge rest/ into network/ and split dto/ into API-layer vs UI-layer,
fulfilling the CLAUDE.md guardrail: all remote-data code now lives under network/.

rest/ dissolved:
- rest/request/ → network/request/ (ComputeRequest, FertilizerRequest, RecommendationRequest, UserInfo)
- rest/response/ → network/response/ (ReverseGeocodeResponse, WeatherResponse)
- rest/ directory deleted

dto/ split:
- API-layer DTOs (Moshi-annotated, .toEntity() methods) → network/dto/
  (Fertilizers, FertilizerPrices, CassavaPrices, CassavaUnits, InvestmentAmounts,
   MaizePrices, StarchFactories, Pagination, RecommendationResponse, UserFeedBackRequest,
   CurrencyDto)
- UI/domain DTOs remain in dto/
  (AdviceCompletionDto, Options, OptionEntries)

Import sites updated: AkilimoApi, NetworkUtils, GeocodingService, LocationIqApi,
WeatherApi, RecommendationBuilder, GetRecommendationViewModel.

Build and all unit tests pass.
