+++ app/src/main/java/com/akilimo/mobile/network/GeocodingService.kt
@@ -3,7 +3,7 @@ package com.akilimo.mobile.network
 import android.content.Context
 import com.akilimo.mobile.network.ApiClient
 import com.akilimo.mobile.network.LocationIqApi
-import com.akilimo.mobile.rest.response.Address
+import com.akilimo.mobile.network.response.Address
 import io.sentry.Sentry
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
