+++ app/src/main/java/com/akilimo/mobile/network/NetworkUtils.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.net.ConnectivityManager
 import android.net.NetworkCapabilities
 import android.os.Build
-import com.akilimo.mobile.dto.ApiErrorResponse
+import com.akilimo.mobile.network.dto.ApiErrorResponse
 import com.squareup.moshi.Moshi
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
