+++ app/src/main/java/com/akilimo/mobile/network/dto/FertilizerPrices.kt
@@ -1,4 +1,4 @@
-package com.akilimo.mobile.dto
+package com.akilimo.mobile.network.dto
 
 import com.akilimo.mobile.entities.FertilizerPrice
 import com.squareup.moshi.Json
